fix the bug

+++ MiddlewareEpsonVision/Form1.cs
@@ -96,7 +96,7 @@ namespace MiddlewareEpsonVision
                     U = float.Parse(values[idx + 3], CultureInfo.InvariantCulture),
                     V = float.Parse(values[idx + 4], CultureInfo.InvariantCulture),
                     W = float.Parse(values[idx + 5], CultureInfo.InvariantCulture),
-                    SprayStatus = int.Parse(values[idx + 6], CultureInfo.InvariantCulture)
+                    SprayStatus = float.Parse(values[idx + 6], CultureInfo.InvariantCulture)
                 });
 
                 UiLogger.Log($"P{i}: X={points[i].X:F3}, Y={points[i].Y:F3}, Z={points[i].Z:F3}, " +
